+++ src/components/[guild]/RequirementCard/hooks/useTokenImage.ts
@@ -6,9 +6,9 @@ const useTokenImage = (address: string): string => {
   const tokens = useTokensList()
   const foundToken = tokens?.find((token) => token.address === address)
 
-  if (foundToken && foundToken.logoURI) return foundToken.logoURI
+  if (!foundToken?.logoURI) return null
 
-  return null
+  return foundToken.logoURI
 }
 
 export default useTokenImage
